More refactoring and cleanup
